fix: `spawn.sync` options in fixer to use `shell: true`
For whatever reason, `spawn.sync` in fixer now requires the `shell: true` option to be set, and the command arg values surrounded in quotes for the phpcbf command to be executed properly.

Not sure whether a change was made in Node, but for documentation purposes and future investigation, this was with Node v22.18.0 and NPM v11.5.2.